% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{check_conditions}
\alias{check_conditions}
\title{Check Conditions for Count Regression}
\usage{
check_conditions(fit)
}
\arguments{
\item{fit}{a fitted model object from fit_count()}
}
\value{
diagnostic results based on model type
}
\description{
Check Conditions for Count Regression
}
