% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEvaluation.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate a Count Regression Model}
\usage{
evaluate_model(fit, fit2 = NULL)
}
\arguments{
\item{fit}{a fitted model object}

\item{fit2}{(optional) a second fitted model for comparison}
}
\value{
a list with model statistics and (optionally) a model comparison
}
\description{
Evaluate a Count Regression Model
}
