% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_count}
\alias{fit_count}
\title{Fit a Count Regression Model}
\usage{
fit_count(data, formula, model = "poisson")
}
\arguments{
\item{data}{data frame}

\item{formula}{formula}

\item{model}{specifying model type}
}
\value{
fitted model
}
\description{
Fit a Count Regression Model
}
