% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{interpret}
\alias{interpret}
\title{Interpret a count regression model}
\usage{
interpret(fit)
}
\arguments{
\item{fit}{a fitted model object}
}
\value{
data frame with rate ratios and interpretations
}
\description{
Interpret a count regression model
}
