% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_count_data}
\alias{summarize_count_data}
\title{Summarize Count Data}
\usage{
summarize_count_data(data, response)
}
\arguments{
\item{data}{data frame}

\item{response}{response variable}
}
\value{
summary statistics
}
\description{
Summarize Count Data
}
